Remove Profiler and remove unused props
